--- apps/shell/src/keplr.d.ts
@@ -1,6 +1,0 @@
-import type { Window as KeplrWindow } from '@keplr-wallet/types';
-
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  interface Window extends KeplrWindow {}
-}
