+++ tsconfig.base.json
@@ -127,5 +127,6 @@
       "@haqq/storybook-host": ["libs/storybook-host/src/index.ts"]
     }
   },
-  "exclude": ["node_modules", "tmp"]
+  "exclude": ["node_modules", "tmp"],
+  "include": ["**/*.d.ts"]
 }
